12/30
開門按鈕修正
