+++ Unity/First version/Assets/Scripts/opendoor_withButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8798d500feae1547b93d89435270051
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
