+++ Unity/First version/Assets/button_opendoor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2d74eb3b370d8b47910dec58bb9c9af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
