+++ Unity/First version/Logs/shadercompiler-AssetImportWorker0.log
@@ -1,4 +1,4 @@
-Base path: 'D:/unity/2020.3.21f1/Editor/Data', plugins path 'D:/unity/2020.3.21f1/Editor/Data/PlaybackEngines'
+Base path: 'D:/Unity/Download/2020.3.21f1/Editor/Data', plugins path 'D:/Unity/Download/2020.3.21f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
 
 Unhandled exception: Protocol error - failed to read magic number (error -2147483644, transferred 0/4)
